Extract "old" template rendering logic
Pulls a large chunk of template processing out of the Executor and into
a separate module. There is still a bit of residue in the Executor,
though.
